unconfirmedDataForm re-format completed
In the model class for PollinatorDataForm, I was simply setting ArrayLists = ArrayLists, which is bad when you clear them since they're simply pointers. The solution was to clone the appropriate arraylists. Since this hasn't come up before, it's unlikely that the issue is present anywhere else in the program, but the ramifications could be bad if it is.